pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coeus-ci"
version = "0.1.0"
description = "Business intelligence OSINT from public data sources"
requires-python = ">=3.10"
license = "AGPL-3.0-only"
dependencies = [
    "click>=8.1",
    "aiohttp>=3.9",
    "rich>=13.0",
    "dnspython>=2.4",
    "python-whois>=0.9",
    "cryptography>=41.0",
    "edgartools>=3.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "aioresponses"]

[project.scripts]
coeus = "coeus.cli:main"

[tool.setuptools.packages.find]
include = ["coeus*"]
